externalize notification names
